backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi[standard]>=0.136.0",
    "httpx>=0.28.1",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
